Move op methods inside `ElemDidDocument` to have access to _keyVault
