+++ common-libs/common/src/services/DidDocumentService/elem-lib/index.ts
@@ -1,2 +1,1 @@
 export * as func from './func'
-export * as op from './op'
